Changed the way of declaration processing (including fix of #339)
In particular this fixed #339 issue with typedef names processing. Also
processing of '..' and '...' was changed considerably. From now '..'
belongs to declaration specifiers while '...' belongs to parameter
declraration instead of beeing separate entity. So some mistakes with
them aren't processed by grammar. For '...' it is checked that it is
located at the end of parameter declaration list.