Detect automatically when print "\n" before #line
In some cases, e.g. for #line directives for structure declarations, it
is not so easy to understand whether a new line is necessary. For
instance, for typedefs on structure declarations it is necesssary.
Otherwise, there will be something like "typedef #line ...". For pure
declarations of structures the new line is not necessary since it was
already printed after the last printed declaration or we are at the
beginning of the file.

The new approach decides whether to print "\n" before #line on the base
of the last printed character automatically.