Update the last printed character just for printing to file
This fixes the previous commit so that there will be no excessive "\n".
Before they may be due to C Backend printer is used not only to produce
C output, but also to convert and store C declarations. In the latter
case before the given commit the last printed character may be set from
that printing rather than from actual C outputting.