Ignore NOP_EXPRs and do not introduce corresponding casts
It is not clear whether corresponding casts are really necessary.
At least differences for test cases for Aspectator demonstrate that
appropriate casts are artificial. Probably more testing will reveal some
unpleasant cases.

Code for treating NOP_EXPRs was added 11 years ago with initial import
of C-Backend. So this change is pretty dangerous. After recent update to
GCC 11 due to this code we get invalid casts that lose "const"
qualifier on some tests. This is caused by changes added to GCC in commit
32934a4f45a7.