Increase expression recursion limit from 10 to 100
This limit is intended to avoid infinite recursion due to unsupported
expressions. Its drawback is that it also can break normal expressions
that have too considerable levels of nesting. Hopefully, most
expressions do not suffer from this, but, say, substitution of macro
roundup_pow_of_two() can hit this limit. This is the case for driver
drivers/net/ethernet/netronome/nfp/nfp.ko of Linux 5.10.27 (probably
this is the only driver that has more than 10 nesting levels).

Nowadays there is just a few expression recursion limits, so it will not
hurt anything substantially if we increase this limit 10 times.